Make a `use` macro for common cache client stuff
